% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{BL_inv_logistic}
\alias{BL_inv_logistic}
\title{Inverse logistic model}
\usage{
BL_inv_logistic(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Parameter describing the possible maximum response.}

\item{beta1}{Scaling parameter.}

\item{beta2}{Parameter describing the shape of the model.}
}
\value{
Numeric vector of response value.
}
\description{
This function fits the inverse logistic model.
}
\examples{
BL_inv_logistic(2, 20, 2, 4)

}
\keyword{internal}
