% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{BL_logisticfm}
\alias{BL_logisticfm}
\title{Fermont logistic model}
\usage{
BL_logisticfm(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Parameter describing the possible maximum response.}

\item{beta1}{Scaling parameter.}

\item{beta2}{Parameter describing the shape of the model.}
}
\value{
Numeric vector of response value.
}
\description{
This function fits the fermont logistic model.
}
\examples{
BL_logisticfm(2, 20, 5, 2)

}
\keyword{internal}
