% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{blm}
\alias{blm}
\title{Linear model}
\usage{
blm(x, beta0, beta1)

blm(x, beta0, beta1)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Intercept parameter.}

\item{beta1}{Slope parameter.}
}
\value{
Numeric vector of response value.
}
\description{
This function fits the linear model.

This function fits the linear model.
}
\examples{
blm(2,10,5)

}
\keyword{internal}
