% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_boundary.R
\name{blqr}
\alias{blqr}
\title{Boundary line model determination using quantile regression}
\usage{
blqr(x,y,model, equation=NULL,theta,tau=0.95,optim.method="Nelder-Mead",
     xmin=min(bound$x),xmax=max(bound$x), plot=TRUE,line_col="red",lwd=1,
     line_smooth=1000,...)
}
\arguments{
\item{x}{A numeric vector of values for the independent variable.}

\item{y}{A numeric vector of values for the response variable.}

\item{model}{Selects the functional form of the boundary line. It includes
\code{"blm"} for linear model, \code{"lp"} for linear plateau model, \code{"mit"}
for the Mitscherlich model, \code{"schmidt"} for the Schmidt model,
\code{"logistic"} for logistic model, \code{"logisticfm"} for logistic model
proposed by Fermont et al (2009), \code{"inv-logistic"} for the inverse logistic
model, \code{"double-logistic"} for the double logistic model, \code{"qd"}
for quadratic model and the \code{"trapezium"} for the trapezium model.}

\item{equation}{A custom model function writen in the form of an R function. Applies
only when argument \code{model="other"}, else it is \code{NULL}.}

\item{theta}{A numeric vector of initial starting values for optimization
in fitting the boundary model. Its length and arrangement depend on the
suggested model: \itemize{
\item For the \code{"blm"} model, it is a vector of length 2 arranged as intercept
and slope.
\item For the \code{"lp"} model, it is a vector of length 3 arranged as intercept,
slope and maximum response.
\item For the \code{"logistic"} and \code{"inv-logistic"} models, it is a
vector of length 3 arranged as the scaling parameter, shape parameter and maximum
response.
\item For the \code{"logisticfm"} model proposed by Fermont et al. (2009), it is a
vector of length 3 arranged as the scaling parameter, shape parameter and maximum
response.
\item For the \code{"double-logistic"} model, it is a vector of length 6 arranged
as the scaling parameter one, shape parameter one, maximum response, maximum
response, scaling parameter two and shape parameter two.
\item For the \code{"qd"} model, it is a vector of length 3 arranged as constant,
linear coefficient and quadratic coefficient.
\item For the \code{"trapezium"} model, it is a vector of length 3  arranged as
intercept one, slope one, maximum response, intercept two and slope two.
\item For the \code{"mit"} model, it is a vector of length 3 arranged as the
intercept, shape parameter and the maximum response.
\item For the \code{"schmidt"} model, it is a vector of length 3 arranged as scaling
parameter, shape parameter (x-value at maximum response ) and maximum response.}}

\item{tau}{The quantile value (0- 1) that represents the boundary
(\code{default is tau = 0.95}).}

\item{optim.method}{Describes the method used to optimize the model as in the
\code{optim()} function. The methods include \code{"Nelder-Mead"}, \code{"BFGS"},
\code{"CG"}, \code{"L-BFGS-B"}, \code{"SANN"} and \code{"Brent"}.}

\item{xmin}{Numeric value that describes the minimum \code{x} value to which the
boundary line is to be fitted (default is \code{min(x)}).}

\item{xmax}{A numeric value that describes the maximum \code{x} value to which the
boundary line is to be fitted (default is \code{max(x)}). \code{xmin} and \code{xmax}
determine the subset of the data set used to fit boundary model.}

\item{plot}{If \code{TRUE}, a plot is part of the output. If \code{FALSE}, plot
is not part of output (default is \code{TRUE}).}

\item{line_col}{Selects the color of the boundary line.}

\item{lwd}{Determines the thickness of the boundary line on the plot (default is 1).}

\item{line_smooth}{Parameter that describes the smoothness of the boundary line.
(default is 1000). The higher the value, the smoother the line.}

\item{...}{Additional graphical parameters.}
}
\value{
A list of length 5 consisting of the fitted model, equation form, parameters
  of the boundary line, the weighted residue sum square. Additionally, a graphical
  representation of the boundary line on the scatter plot is produced.
}
\description{
This function fits a boundary model to the upper bounds of a scatter plot of
\code{x} and \code{y} by estimating the conditional quantile (0-1) of the
response variable, \code{y}, across values of the predictor variables, \code{x}.
This is achieved using optimization procedure and hence requires some starting
guess parameters of a proposed model.
}
\details{
Some inbuilt models are available for the \code{blqr()} function. The suggest model
forms are as follows: \enumerate{
 \item Linear model (\code{"blm"})
 \deqn{y=\beta_1 + \beta_2x}
 where \eqn{\beta_1} is the intercept and \eqn{\beta_2} is the slope.

 \item Linear plateau model (\code{"lp"})
 \deqn{y= {\rm min}(\beta_1+\beta_2x, \beta_0)}
 where \eqn{\beta_1} is the intercept , \eqn{\beta_2} is the slope  and \eqn{\beta_0}
 is the maximum response.

 \item The logistic (\code{"logistic"}) and inverse logistic (\code{"inv-logistic"})
 models
 \deqn{ y= \frac{\beta_0}{1+e^{\beta_2(\beta_1-x)}}}
 \deqn{ y= \beta_0 - \frac{\beta_0}{1+e^{\beta_2(\beta_1-x)}}}
 where \eqn{\beta_1} is a scaling parameter , \eqn{\beta_2} is a shape parameter
 and \eqn{\beta_0} is the maximum response.

 \item Logistic model (\code{"logisticfm"})  (Fermont et al. 2009)
 \deqn{ y= \frac{\beta_0}{1+(\beta_1 \times e^{-\beta_2x})}}
  where \eqn{\beta_1} is a scaling parameter, \eqn{\beta_2} is a shape
  parameter and \eqn{\beta_0} is the maximum response.

 \item Double logistic model (\code{"double-logistic"})
 \deqn{ y= \frac{\beta_{0,1}}{1+e^{\beta_2(\beta_1-x)}} -
 \frac{\beta_{0,2}}{1+e^{\beta_4(\beta_3-x)}}}
 where \eqn{\beta_1} is a scaling parameter one, \eqn{\beta_2} is a shape parameter one,
 \eqn{\beta_{0,1}} and \eqn{\beta_{0,2}} are the maximum response ,
 \eqn{\beta_3} is a scaling parameter two and  \eqn{\beta_4} is a shape parameter two.

 \item Quadratic model (\code{"qd"})
 \deqn{y=\beta_1 + \beta_2x + \beta_3x^2}
 where \eqn{\beta_1} is a constant, \eqn{\beta_2} is a linear coefficient
 and  \eqn{\beta_3} is the quadratic coefficient.

 \item Trapezium model (\code{"trapezium"})
 \deqn{y={\rm min}(\beta_1+\beta_2x, \beta_0, \beta_3 + \beta_4x)}
 where  \eqn{\beta_1} is the intercept one, \eqn{\beta_2} is the slope one,
 \eqn{\beta_0} is the maximum response, \eqn{\beta_3} is the intercept two
 and \eqn{\beta_3} is the slope two.

 \item Mitscherlich model (\code{"mit"})
 \deqn{y= \beta_0 + \beta_1(1-e^{\frac{-x}{\beta_2}})}
 where \eqn{\beta_1} is the intercept, \eqn{\beta_2} is a shape parameter
 and \eqn{\beta_0} is the maximum response.

 \item Schmidt model (\code{"schmidt"})
 \deqn{y= \beta_0 + \beta_1(x-\beta_2)^2}
 where \eqn{\beta_1} is ascaling parameter, \eqn{\beta_2} is a
 shape parameter (x-value at maximum response ) and \eqn{\beta_0} is the
 maximum response .

 \item Custom model ("other")
 This option allows you to create your own model form using the function
 \code{function()}. The custom model should be assigned to the argument
 \code{equation}. Note that the parameters for the custom model should be
 \code{a} and \code{b} for a two parameter model; \code{a}, \code{b} and \code{c}
 for a three parameter model; \code{a}, \code{b}, \code{c} and \code{d} for a
 four parameter model and so on.
 }

The function \code{blbin()} utilities the optimization procedure of the
\code{optim()} function to determine the model parameters. There is a tendency
for optimization algorithms to settle at a local optimum. To remove the risk of
settling for local optimum parameters, it is advised that the function is run
using several starting values and the results with the smallest error
(weighted residue sum square) can be taken as a representation of the global
 optimum.
}
\examples{

x<-evapotranspiration$`ET(mm)`
y<-evapotranspiration$`yield(t/ha)`

blqr(x,y,theta = c(0.5,0.02), model = "blm", tau = 0.95)

}
\references{
Cade, B. S., & Noon, B. R. (2003). A gentle introduction to quantile regression
for ecologists. Frontiers in Ecology and the Environment, 1(8), 412-420.

Dhanoa, M. S., Sanderson, R., Cardenas, L. M., Shepherd, A., Chadwick, D. R.,
Powell, C. D., ... & France, J. (2022). Overview and application of the Mitscherlich
equation and its extensions to estimate the soil nitrogen pool fraction associated
with crop yield and nitrous oxide emission. Advances in Agronomy, 174, 269-295.

Fermont, A. M., Van Asten, P. J., Tittonell, P., Van Wijk, M. T., &
Giller, K. E. (2009).Closing the cassava yield gap: an analysis from smallholder
farms in East Africa. Field Crops Research, 112 (1), 24–36.

Schmidt, U., Thöni, H., & Kaupenjohann, M. (2000). Using a boundary line approach
to analyze N2O flux data from agricultural soils. Nutrient Cycling in Agroecosystems,
57, 119-129.
}
\author{
Chawezi Miti <chawezi.miti@nottingham.ac.uk>
}
