% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_and_load_tcltk}
\alias{check_and_load_tcltk}
\title{Check and Load Tcl/Tk}
\usage{
check_and_load_tcltk()
}
\value{
Logical value indicating whether `tcltk` was successfully loaded.
}
\description{
This function checks if the Tcl/Tk environment is available and loads the `tcltk`
package if possible.
It returns TRUE if `tcltk` is successfully loaded, otherwise FALSE.
}
\keyword{internal}
