% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{coffcturb}
\alias{coffcturb}
\title{Coffcturb}
\usage{
coffcturb(x, mu, sig, a, c, sigh)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{mu}{Mean of x.}

\item{sig}{Fitting parameter.}

\item{a}{Fitting parameter.}

\item{c}{Fitting parameter.}

\item{sigh}{Measurement error.}
}
\value{
Parameters input for the coffcturb function.
}
\description{
This function determines the coffcturb parameters.
}
\keyword{internal}
