% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{coffcturb3}
\alias{coffcturb3}
\title{Coffcturb}
\usage{
coffcturb3(x, mu, sig, a, c, sigh)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{mu}{Mean of x.}

\item{sig}{Fitting parameter.}

\item{a}{Fitting parameter.}

\item{c}{Fitting parameter.}

\item{sigh}{Measurement error.}
}
\value{
Parameters of coffcturb.
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
