% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{coffcturb4}
\alias{coffcturb4}
\title{Coffcturb}
\usage{
coffcturb4(x, mu, sig, a, c, sigh)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{mu}{Mean of x.}

\item{sig}{Fitting parameter.}

\item{a}{Fitting parameter.}

\item{c}{Fitting parameter.}

\item{sigh}{Measurement error.}
}
\value{
Parameters of coffcturb.
}
\description{
This function determines the parameters of  coffcturb function.
}
\keyword{internal}
