% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{drawBL}
\alias{drawBL}
\title{Draw the boundary line}
\usage{
drawBL(x, beta0, beta1, beta2, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Model parameter.}

\item{beta1}{Model parameter.}

\item{beta2}{Model parameter.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Draws the model on plot.
}
\description{
This function draws the boundary line on the plot.
}
\keyword{internal}
