% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{drawBL2}
\alias{drawBL2}
\title{Draws boundary line}
\usage{
drawBL2(x, beta0, beta1, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Intercept parameter.}

\item{beta1}{Slope parameter.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Draws model line on plot.
}
\description{
This function draws the boundary line on a plot.
}
\keyword{internal}
