% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{drawBL3}
\alias{drawBL3}
\title{Draw trapezium model}
\usage{
drawBL3(x, beta0, beta1, beta2, beta3, beta4, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Model parameter.}

\item{beta1}{Model parameter.}

\item{beta2}{Model parameter.}

\item{beta3}{Model parameter.}

\item{beta4}{Model parameter.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Draws the model on plot.
}
\description{
This function draws the trapezium model on a plot.
}
\keyword{internal}
