% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{drawBL4}
\alias{drawBL4}
\title{Draw double logistic model}
\usage{
drawBL4(x, beta01, beta02, beta1, beta2, beta3, beta4, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta01}{Model parameter.}

\item{beta02}{Model parameter.}

\item{beta1}{Model parameter.}

\item{beta2}{Model parameter.}

\item{beta3}{Model parameter.}

\item{beta4}{Model parameter.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Draws the model on plot.
}
\description{
This function draws the double logistic boundary line on a plot.
}
\keyword{internal}
