% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdens_maxyield3}
\alias{jdens_maxyield3}
\title{Joint density of observed values}
\usage{
jdens_maxyield3(X, ymax, sigh, mux, muy, sdx, sdy, rcorr)
}
\arguments{
\item{X}{Numeric vector of independent variable.}

\item{ymax}{Maximum response.}

\item{sigh}{Measurement error.}

\item{mux}{Mean of x.}

\item{muy}{Mean of y.}

\item{sdx}{Standard deviation of x.}

\item{sdy}{Standard deviation of y.}

\item{rcorr}{Correlation of x and y.}
}
\value{
Joint density of observed values x and y.
}
\description{
This function determines the joint density of observed values x and y.
}
\keyword{internal}
