% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdens_maxyield4}
\alias{jdens_maxyield4}
\title{Joint density of observed values}
\usage{
jdens_maxyield4(X, ymax, sigh, mux, muy, sdx, sdy, rcorr)
}
\arguments{
\item{X}{Numeric vector of independent variable.}

\item{ymax}{maximum response.}

\item{sigh}{Measurement error.}

\item{mux}{mean of x.}

\item{muy}{mean of y.}

\item{sdx}{standard deviation of x.}

\item{sdy}{standard deviation of y.}

\item{rcorr}{correlation of x and y.}
}
\value{
Joint density of observed values x and y.
}
\description{
This function determines the joint density of observed values.
}
\keyword{internal}
