% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdensup}
\alias{jdensup}
\title{Joint density of observed values}
\usage{
jdensup(X, BLMod, beta0, beta1, beta2, sigh, mux, muy, sdx, sdy, rcorr)
}
\arguments{
\item{X}{Numeric vector containing x and y values.}

\item{BLMod}{Determines the boundary line model fitted.}

\item{beta0}{Model parameter.}

\item{beta1}{Model parameter.}

\item{beta2}{Model parameter.}

\item{sigh}{Measurement error.}

\item{mux}{Mean of x.}

\item{muy}{Mean of y.}

\item{sdx}{Standard deviation of x.}

\item{sdy}{Standard deviation of y.}

\item{rcorr}{Correlation of x and y.}
}
\value{
Parameters of jdensup.
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
