% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdensup2}
\alias{jdensup2}
\title{Joint density of observed}
\usage{
jdensup2(X, BLMod, beta0, beta1, sigh, mux, muy, sdx, sdy, rcorr)
}
\arguments{
\item{X}{Numeric vector of independent variable.}

\item{BLMod}{Determines the boundary line model fitted.}

\item{beta0}{Model parameter.}

\item{beta1}{Model parameter.}

\item{sigh}{Measurement error.}

\item{mux}{Mean of x.}

\item{muy}{Mean of y.}

\item{sdx}{Standard deviation of x.}

\item{sdy}{Standard deviation of y.}

\item{rcorr}{Correlation of x and y.}
}
\value{
Joint density of observed values x and y.
}
\description{
This function determines the joint density of observed values x and y given
beta0,beta1,beta2 as boundary line parameters.
}
\keyword{internal}
