% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdensup4}
\alias{jdensup4}
\title{Internal functions}
\usage{
jdensup4(
  X,
  BLMod,
  beta01,
  beta02,
  beta1,
  beta2,
  beta3,
  beta4,
  sigh,
  mux,
  muy,
  sdx,
  sdy,
  rcorr
)
}
\arguments{
\item{X}{input in function}

\item{BLMod}{determines the boundary line model fitted}

\item{beta01}{model parameter}

\item{beta02}{model parameter}

\item{beta1}{model parameter}

\item{beta2}{model parameter}

\item{beta3}{model parameter}

\item{beta4}{model parameter}

\item{sigh}{measurement error}

\item{mux}{mean of x}

\item{muy}{mean of y}

\item{sdx}{standard deviation of x}

\item{sdy}{standard deviation of y}

\item{rcorr}{correlation of x and y}
}
\value{
Parameters of jdensup
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
