% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurt.R
\name{kurt}
\alias{kurt}
\title{kurtosis}
\usage{
kurt(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
The reduced coefficient of kurtosis.
}
\description{
A function to calculate an estimate of the coefficient of kurtosis
from a set of data.
}
\examples{
x<-evapotranspiration$`ET(mm)`
kurt(x)

}
\author{
Richard Murray Lark <murray.lark@nottingham.ac.uk>
}
\keyword{internal}
