% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{lp}
\alias{lp}
\title{Linear plateau model}
\usage{
lp(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Plateau parameter.}

\item{beta1}{Slope parameter.}

\item{beta2}{Intercept parameter.}
}
\value{
Numeric vector of response value.
}
\description{
This is function fits the linear plateau model.
}
\examples{
lp(2, 20, 5, 2)

}
\keyword{internal}
