% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{mit}
\alias{mit}
\title{Internal functions}
\usage{
mit(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{independent variable}

\item{beta0}{model parameter}

\item{beta1}{model parameter}

\item{beta2}{model parameter}
}
\value{
Parameters of the model.
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
