% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_drop.R
\name{na.drop}
\alias{na.drop}
\title{Remove NA values}
\usage{
na.drop(xin)
}
\arguments{
\item{xin}{A numeric vector.}
}
\value{
A vector without missing values
}
\description{
Removes missing values from a dataset.
}
\examples{
x<-evapotranspiration$`ET(mm)`
na.drop(x)

}
\author{
Richard Murray Lark <murray.lark@nottingham.ac.uk>
}
\keyword{internal}
