% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef}
\alias{nll_mef}
\title{Internal function}
\usage{
nll_mef(pars, uplo, BLMod)
}
\arguments{
\item{pars}{Object from  jdensup function.}

\item{uplo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
The nll for 3-parameter BL model.
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
