% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef2}
\alias{nll_mef2}
\title{Internal functions}
\usage{
nll_mef2(pars, uplo, BLMod)
}
\arguments{
\item{pars}{input in nll_mef function}

\item{uplo}{determines whether it is upper or lower boundary}

\item{BLMod}{determines the boundary line model fitted}
}
\value{
Parameters nll_mef2 for linear model
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
