% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef3}
\alias{nll_mef3}
\title{The nll for 5-parameter BL model}
\usage{
nll_mef3(pars, uplo, BLMod)
}
\arguments{
\item{pars}{Object from  jdensup function.}

\item{uplo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Parameters of nll model.
}
\description{
This function determines the nll for 5-parameter BL models.
}
\keyword{internal}
