% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef4}
\alias{nll_mef4}
\title{The nll for 6-parameter BL model}
\usage{
nll_mef4(pars, uplo, BLMod)
}
\arguments{
\item{pars}{Object from  jdensup function.}

\item{uplo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Parameters of nll_mef.
}
\description{
This function determines the nll for a 6-parameter BL model.
}
\keyword{internal}
