% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef_maxyield}
\alias{nll_mef_maxyield}
\title{Internal functions}
\usage{
nll_mef_maxyield(pars)
}
\arguments{
\item{pars}{input in nll_mef function}
}
\value{
Parameters of the max response model
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
