% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef_maxyield4}
\alias{nll_mef_maxyield4}
\title{Flat upper BL model}
\usage{
nll_mef_maxyield4(pars)
}
\arguments{
\item{pars}{Object from  jdensup function.}
}
\value{
Parameters of the max response model.
}
\description{
This function returns nll for simple flat upper BL model.
}
\keyword{internal}
