% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nllmvn3}
\alias{nllmvn3}
\title{Multivariate normal model}
\usage{
nllmvn3(pars)
}
\arguments{
\item{pars}{Object from  jdensup function.}
}
\value{
Parameters of null multivariate normal model.
}
\description{
This function determines the parameters of the multivariate normal model.
}
\keyword{internal}
