% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nllmvn4}
\alias{nllmvn4}
\title{Multivariate normal model}
\usage{
nllmvn4(pars)
}
\arguments{
\item{pars}{Object from  jdensup function.}
}
\value{
Parameters of null model.
}
\description{
This function determines parameters of the bivariate normal model.
}
\keyword{internal}
