% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octile.R
\name{ocskew}
\alias{ocskew}
\title{Octile Skewness}
\usage{
ocskew(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
The octile skewness.
}
\description{
A function to calculate an estimate of the octile skewness from a set of data.
}
\examples{
x<-evapotranspiration$`ET(mm)`
ocskew(x)

}
\author{
Richard Murray Lark <murray.lark@nottingham.ac.uk>
}
\keyword{internal}
