% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{par_nll_mef3}
\alias{par_nll_mef3}
\title{Partial derivative}
\usage{
par_nll_mef3(x, UpLo, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{UpLo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Rough partial derivative at x of nll.
}
\description{
This function determines the rough partial derivative at x of nll.
}
\keyword{internal}
