% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{par_nll_mef4}
\alias{par_nll_mef4}
\title{Partial derivative}
\usage{
par_nll_mef4(x, UpLo, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{UpLo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Rough partial derivative at x.
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
