% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printcm_class.R
\name{print.cm}
\alias{print.cm}
\title{Print class cm}
\usage{
\method{print}{cm}(x, ...)
}
\arguments{
\item{x}{Print object.}

\item{...}{Other parameters associated with the \code{print()} function.}
}
\value{
A object containing only the first four items.
}
\description{
This is an S3 print function that Prints only the first 4 elements of
cm class objects.
}
\examples{
numbers<- 1:10
class(numbers)<-"cm"
numbers

}
\keyword{internal}
