% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printwm_class.R
\name{print.wm}
\alias{print.wm}
\title{Print class wm}
\usage{
\method{print}{wm}(x, ...)
}
\arguments{
\item{x}{Print object.}

\item{...}{Other parameters associated with the \code{print()} function.}
}
\description{
Prints an output with 5 elements.
}
\keyword{internal}
