% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{qd}
\alias{qd}
\title{Quadratic functions}
\usage{
qd(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Parameter describing the constant.}

\item{beta1}{Linear parameter.}

\item{beta2}{Quadratic parameter.}
}
\value{
Numeric vector of response value.
}
\description{
This is function fits the quadratic model.
}
\examples{
qd(2, 0.5,2,1)

}
\keyword{internal}
