% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{schmidt}
\alias{schmidt}
\title{Schmidt model}
\usage{
schmidt(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Parameter describing the possible maximum response.}

\item{beta1}{Value of x at which the response is maximum.}

\item{beta2}{A scaling parameter, which is zero if the boundary is a constant.}
}
\value{
Numeric vector of response value.
}
\description{
This is function fits the Schmidt model.
}
\examples{
schmidt(2, 20, 5, 2)

}
\keyword{internal}
