% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{seHessian}
\alias{seHessian}
\title{Internal functions}
\usage{
seHessian(a, hessian = FALSE, silent = FALSE)
}
\arguments{
\item{a}{determines the boundary line model fitted}

\item{hessian}{If true hessian is used}

\item{silent}{condtion of matrix}
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
