% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Skewness}
\usage{
skew(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
The coefficient of skewness.
}
\description{
Function to calculate an estimate of the coefficient of skewness from
a set of data.
}
\examples{
x<-evapotranspiration$`ET(mm)`
skew(x)

}
\author{
Richard Murray Lark <murray.lark@nottingham.ac.uk>
}
\keyword{internal}
