% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soil}
\alias{soil}
\title{Soil survey data}
\format{
A data.frame with 6110 rows and 3 columns:
\describe{
\item{yield}{ Wheat yield (ton/ha) measured}.
\item{pH}{ pH measurement}.
\item{P}{ soil p (mg/kg)}.
}
}
\usage{
soil
}
\description{
This data set is a subset of a data set that was assembled by AgSpace Agriculture Ltd
in a soil survey conducted on farms in various management units across England. The survey
included measurements of wheat yield as well as various soil parameters. This
particular dataset only contains the soil properties pH and  phosphorus (P).
}
\details{
This data set should only be used for illustration purposes for this package and should not be used
in any form publication without permission from the owners.
}
\examples{
data(soil)
}
\keyword{datasets}
