% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilP.R
\docType{data}
\name{SoilP}
\alias{SoilP}
\title{Soil Phosphorus data}
\format{
A data.frame with 6020 rows and 2 columns:
\describe{
\item{yield}{ Wheat yield (ton/ha) measured}.
\item{P}{ Phosphorus (mg/l)}.
}
}
\usage{
SoilP
}
\description{
This data set is a subset of a data set that was assembled by AgSpace Agriculture Ltd
in a soil survey conducted on farms in various management units across England.
}
\examples{
data(SoilP)
}
\keyword{datasets}
