% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilpH.R
\docType{data}
\name{SoilpH}
\alias{SoilpH}
\title{Soil pH data}
\format{
A data.frame with 6047 rows and 2 columns:
\describe{
\item{yield}{ Wheat yield (ton/ha) measured}.
\item{pH}{ pH measurement}.
}
}
\usage{
SoilpH
}
\description{
This data set is a subset of a data set that was assembled by AgSpace Agriculture Ltd
in a soil survey conducted on farms in various management units across England.
}
\examples{
data(SoilpH)
}
\keyword{datasets}
