% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartValues.R
\name{startValues}
\alias{startValues}
\title{Starting values for optimization functions}
\usage{
startValues(model = "blm", p = NULL, digits = 2, ...)
}
\arguments{
\item{model}{Selects the functional form of the boundary line. It includes
\code{"blm"} for linear model, \code{"lp"} for
linear plateau model, \code{"mit"} for the Mitscherlich model, \code{"schmidt"}
for the Schmidt model, \code{"logistic"} for logistic model, \code{"logisticfm"}
for logistic model proposed by Fermont et al (2009), \code{"inv-logistic"} for
the inverse logistic model, \code{"double-logistic"} for the double logistic model,
\code{"qd"} for quadratic model and the \code{"trapezium"} for the trapezium model.
The default is \code{"blm"}.}

\item{p}{The number of selected points used to obtain start values for the logistic
type models. It is \code{NULL} for other models.}

\item{digits}{Number of decimal points for logistic type models (default is 2).}

\item{...}{Additional graphical parameters. Applies to the logistic models to
control the text on the plot.}
}
\value{
A list containing the parameters of the suggested model.
}
\description{
This functions helps to determine initial values for a selected boundary line
model when using the functions \code{blbin()}, \code{blqr()}, \code{BOLIDES()},
\code{cbvn()} and \code{ble_profile()} to determine model parameters.
}
\details{
This function uses the \code{locator()} function. Once the model is selected,
the points that make up the boundary points are selected using mouse click on
the plots.
}
\author{
Chawezi Miti <chawezi.miti@nottingham.ac.uk>
}
