% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summa.R
\name{summa}
\alias{summa}
\title{Summary statistics}
\usage{
summa(x, sigf)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{sigf}{The number of significant figures to report (optional).}
}
\value{
A matrix containing the mean value, median value,
  first and third quartiles, sample variance, sample standard deviation,
  coefficient of skewness, octile skewness, coefficient of kurtosis and
  the number of probable outliers in a data set.
}
\description{
A function to calculate summary statistics of a set of data.
}
\examples{
x<-evapotranspiration$`ET(mm)`
summa(x,2)
}
\author{
Richard Murray Lark <murray.lark@nottingham.ac.uk>
}
