% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaplot.R
\name{summaplot}
\alias{summaplot}
\title{Summary plots}
\usage{
summaplot(x, varname)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{varname}{The name of the variable (optional), character so in quotes e.g.
"Clay content". If not used then the variable is called x on plots.}
}
\value{
A histogram with a boxplot over it and QQ plot of the variable x.
}
\description{
A function to produce summary plots of a set of data.
}
\examples{

x<-evapotranspiration$`ET(mm)`
summaplot(x)

}
\author{
Richard Murray Lark <murray.lark@nottingham.ac.uk>
}
