% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{BL_logisticfm}
\alias{BL_logisticfm}
\title{Internal functions}
\usage{
BL_logisticfm(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{independent variable}

\item{beta0}{is a parameter describing the possible maximum response}

\item{beta1}{is a scaling parameter}

\item{beta2}{is a parameter describing the shape of the model}
}
\value{
Parameters of the model.
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
