% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{blm}
\alias{blm}
\title{Internal functions}
\usage{
blm(x, beta0, beta1)

blm(x, beta0, beta1)
}
\arguments{
\item{x}{independent variable}

\item{beta0}{model parameter}

\item{beta1}{model parameter}
}
\value{
Parameters of the model

Parameters of BL_lm.
}
\description{
This is a set of functions that support the censored bivariate normal model.

This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
