% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{coffcturb3}
\alias{coffcturb3}
\title{Internal functions}
\usage{
coffcturb3(x, mu, sig, a, c, sigh)
}
\arguments{
\item{x}{independent variable}

\item{mu}{mean}

\item{sig}{fitting parameter}

\item{a}{fitting parameter}

\item{c}{fitting parameter}

\item{sigh}{measurement error}
}
\value{
Parameters of coffcturb
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
