% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{drawBL}
\alias{drawBL}
\title{Internal functions}
\usage{
drawBL(x, beta0, beta1, beta2, BLMod)
}
\arguments{
\item{x}{independent variable}

\item{beta0}{model parameter}

\item{beta1}{model parameter}

\item{beta2}{model parameter}

\item{BLMod}{determines the boundary line model fitted}
}
\value{
draws the model on plot
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
