% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdens_maxyield}
\alias{jdens_maxyield}
\title{Internal functions}
\usage{
jdens_maxyield(X, ymax, sigh, mux, muy, sdx, sdy, rcorr)
}
\arguments{
\item{X}{input in function}

\item{ymax}{maximumresponse}

\item{sigh}{measurement error}

\item{mux}{mean of x}

\item{muy}{mean of y}

\item{sdx}{standard deviation of x}

\item{sdy}{standard deviation of y}

\item{rcorr}{correlation of x and y}
}
\value{
Parameters of jdens_maxyield
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
