% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{jdensup2}
\alias{jdensup2}
\title{Internal functions}
\usage{
jdensup2(X, BLMod, beta0, beta1, sigh, mux, muy, sdx, sdy, rcorr)
}
\arguments{
\item{X}{input in function}

\item{BLMod}{determines the boundary line model fitted}

\item{beta0}{model parameter}

\item{beta1}{model parameter}

\item{sigh}{measurement error}

\item{mux}{mean of x}

\item{muy}{mean of y}

\item{sdx}{standard deviation of x}

\item{sdy}{standard deviation of y}

\item{rcorr}{correlation of x and y}
}
\value{
Parameters of jdensup2
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
