% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{mit}
\alias{mit}
\title{Mitscherlich function}
\usage{
mit(x, beta0, beta1, beta2)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Scaling parameter.}

\item{beta1}{Model parameter maximum response.}

\item{beta2}{Shape parameter.}
}
\value{
Numeric vector of response value.
}
\description{
This is function fits the mitscherlich model.
}
\examples{
mit(2, 20,1,5)

}
\keyword{internal}
