% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef2}
\alias{nll_mef2}
\title{Internal functions}
\usage{
nll_mef2(pars, uplo, BLMod)
}
\arguments{
\item{pars}{Object from  jdensup function.}

\item{uplo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
The nll for 2-parameter BL model.
}
\description{
This function produces nll for 3-parameter BL model.
}
\keyword{internal}
