% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nll_mef4}
\alias{nll_mef4}
\title{Internal functions}
\usage{
nll_mef4(pars, uplo, BLMod)
}
\arguments{
\item{pars}{input in nll_mef function}

\item{uplo}{determines whether it is upper or lower boundary}

\item{BLMod}{determines the boundary line model fitted}
}
\value{
Parameters of nll_mef
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\author{
Chawezi Miti <chawezi.miti@nottingham.ac.uk>
}
\keyword{internal}
