% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nllmvn}
\alias{nllmvn}
\title{Multivariate normal}
\usage{
nllmvn(pars)
}
\arguments{
\item{pars}{Object from  jdensup function.}
}
\value{
Parameters of null model.
}
\description{
This function fits the multivariate normal model.
}
\keyword{internal}
