% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{nllmvn3}
\alias{nllmvn3}
\title{Internal functions}
\usage{
nllmvn3(pars)
}
\arguments{
\item{pars}{input in nll_mef function}
}
\value{
Parameters of null model
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
