% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{par_nll_mef}
\alias{par_nll_mef}
\title{Internal functions}
\usage{
par_nll_mef(x, UpLo, BLMod)
}
\arguments{
\item{x}{input in nll_mef function}

\item{UpLo}{determines whether it is upper or lower boundary}

\item{BLMod}{determines the boundary line model fitted}
}
\value{
Parameters of par_nll_mef
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\keyword{internal}
