% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{par_nll_mef2}
\alias{par_nll_mef2}
\title{Partial derivative}
\usage{
par_nll_mef2(x, UpLo, BLMod)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{UpLo}{Determines whether it is upper or lower boundary.}

\item{BLMod}{Determines the boundary line model fitted.}
}
\value{
Parameters of par_nll_mef2
}
\description{
This function determines rough partial derivative of x.
}
\keyword{internal}
