% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printcm_class.R
\name{print.cm}
\alias{print.cm}
\title{Print class cm}
\usage{
\method{print}{cm}(x, ...)
}
\arguments{
\item{x}{Print object.}

\item{...}{Other parameters associated with the \code{print()} function.}
}
\description{
Prints an output with 5 elements.
}
\keyword{internal}
