% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printwm_class.R
\name{print.wm}
\alias{print.wm}
\title{Print class wm}
\usage{
\method{print}{wm}(x, ...)
}
\arguments{
\item{x}{Print object.}

\item{...}{Other parameters associated with the \code{print()} function.}
}
\value{
A object containing only the first four items.
}
\description{
This is an S3 print function that Prints only the first 4 elements of
wm class objects.
}
\examples{
numbers<- 1:10
class(numbers)<-"wm"
numbers

}
\keyword{internal}
