% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeHessian.R
\name{seHessian}
\alias{seHessian}
\title{Hessian matrix}
\usage{
seHessian(a, hessian = FALSE, silent = FALSE)
}
\arguments{
\item{a}{The hessian matrix.}

\item{hessian}{If `True`, hessian matrix is used.}

\item{silent}{Condition of matrix.}
}
\value{
The standard error from hessian matrix
}
\description{
This is a set of functions that support the censored bivariate normal model.
}
\examples{

hessian<-matrix(c(37.45965, 83.0686,83.06863,188.92427),2,2)

seHessian(hessian) # calculates the standard error

}
\keyword{internal}
