% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_functions.R
\name{trapezium}
\alias{trapezium}
\title{Trapezium model}
\usage{
trapezium(x, beta0, beta1, beta2, beta3, beta4)
}
\arguments{
\item{x}{Numeric vector of independent variable.}

\item{beta0}{Maximum response model parameter.}

\item{beta1}{Intercept one model parameter.}

\item{beta2}{Slope one model parameter.}

\item{beta3}{Intercept two model parameter.}

\item{beta4}{Slope two model parameter.}
}
\value{
A numeric vector of response values.
}
\description{
This function fits the trapezium model.
}
\examples{
trapezium(3, 20, 5, 2, 5,2)

}
\keyword{internal}
